resolved handleLogout edits
