--- packages/ui/.env.development
@@ -1,1 +1,0 @@
-IS_NETWORK_MAINNET="true";
